FlaskApp/FlaskApp/requirements.txt
Flask
gunicorn
SQLAlchemy
psycopg2-binary
python-dotenv
requests
pytz
tzdata